Add files via upload
Here we had data of airbnb and the reviews provided to the hotels. We can do the data cleaning for the missing values, work on outliers present for the minimum_nights column. We need to understand minimum nights concept in hotels and stays for guests to work on that outlier.  However, we can also leave the outliers as they are as most of them are for Entire House Property( Room type feature) 

Please go through columns to understand the description. Everything else has been mentioned in the Jupyter Notebook